upcode fix role & createUser

+++ resources/views/admin/category/row.blade.php
@@ -50,22 +50,24 @@
                     <div>Sửa</div>
                 </a>
             </div>
-            <div class="icon-wrapper iconAction">
-                <a href="{{ route('admin.'.$typeRoute.'.view', ['id' => $item->id, 'language' => 'vi', 'type' => 'copy']) }}">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-copy"><rect x="9" y="9" width="13" height="13" rx="2" ry="2"></rect><path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"></path></svg>
-                    <div>Chép</div>
-                </a>
-            </div>
-            {{-- <div class="icon-wrapper iconAction">
-                <div class="actionDelete" onclick="deleteItem({{ $item->id }});">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-x-square">
-                        <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
-                        <line x1="9" y1="9" x2="15" y2="15"></line>
-                        <line x1="15" y1="9" x2="9" y2="15"></line>
-                    </svg>
-                    <div>Xóa</div>
+            @if(auth()->user()->hasRole('admin'))
+                <div class="icon-wrapper iconAction">
+                    <a href="{{ route('admin.'.$typeRoute.'.view', ['id' => $item->id, 'language' => 'vi', 'type' => 'copy']) }}">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-copy"><rect x="9" y="9" width="13" height="13" rx="2" ry="2"></rect><path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"></path></svg>
+                        <div>Chép</div>
+                    </a>
                 </div>
-            </div> --}}
+                <div class="icon-wrapper iconAction">
+                    <div class="actionDelete" onclick="deleteItem({{ $item->id }});">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-x-square">
+                            <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
+                            <line x1="9" y1="9" x2="15" y2="15"></line>
+                            <line x1="15" y1="9" x2="9" y2="15"></line>
+                        </svg>
+                        <div>Xóa</div>
+                    </div>
+                </div>
+            @endif
         </td>
     </tr>
     {{-- <tr id="oneItemSub-{{ $item->id }}">
